Add `languageName` function to core
Adds two new functions:
* `input.languageName()` on an expression will look the input up as an ISO language code, and find the name of the language in the JVM's default locale
* `input.languageName("en")` on an expression would look the input up as an ISO language code, and find the name of the language in the described target locale

Both functions are in `core` since they just rely on JVM internals, and are available in `BuiltinFunctionResolver`
Tests have been added in `LanguageNameSuite.json`

Note that `java.util.Locale` can just return the input if it doesn't recognise it, so we intercept that and turn it into a `MissingField`.
If you'd like that behaviour please use `(input.languageName() | input)` to fall back to the input.